django-app/requirements.txt
django==4.2.0
djangorestframework==3.14.0
pypdf==3.1.0
pyyaml==5.4.0
requests==2.28.0
